PWN/Paging/requirements.txt
capstone
construct
unicorn
keystone-engine
quantumrandom
pycrypto
